+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-invoice
 
+## 6.1.0 (IN PROGRESS)
+
+* Add search index for and export Folio Invoice Number. Refs UINV-556.
+
 ## [6.0.3](https://github.com/folio-org/ui-invoice/tree/v6.0.3) (2024-04-19)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v6.0.2...v6.0.3)